Update new_vis_styles.py
Now able to take downloaded data from the RIT database. Visualizations don't look very good because their strain data is in the spinning frame of the BBHs, so that needs to be adjusted. Also the RIT sonification function is unfinished, since it will have to pull frequency from the strain data instead of horizons - no trajectory or AH data :(