Update new_vis_styles.py
Almost done with functionality to show approximate BH positions and horizons for RIT data. Uses l=2, m=2 mode to get the phase of the BH's orbit, and an ODE involving their orbital velocity to find their separation distance. COM should be (0, 0, 0), and center positions are assumed to be in the x-y plane.